// src/infrastructure/external-services/ai-generate-image/comfyui/control-net/types/NodeControlnetEnum.ts
export enum NodeControlnetEnum {
  CONTROLNET_LOADER = '%%controlnet_loader%%',
  CONTROLNET_LOADER_CONTROLNET_NAME = '%%controlnet_loader_controlnet_net_name%%',
  CONTROLNET_LOAD_IMAGE = '%%controlnet_load_image%%',
  CONTROLNET_LOAD_IMAGE_IMAGE = '%%controlnet_load_image_image%%',
  CONTROLNET_APPLY = '%%controlnet_apply%%',
  POSITIVE_PROMPT = '%%positive_prompt%%',
  CONTROLNET_AIO_PREPROCESSOR_ID = '%%controlnet_aio_preprocessor%%',
  CONTROLNET_AIO_TYPE_PREPROCESSOR = '%%type_preprocessor%%',
}